+++ examples/cpp/redis/include/e2_pipeline.cc
@@ -1,5 +1,5 @@
 #include "common/config.h"
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-kernel/core/runtime.h>
 #include <algorithm>
 #include <array>