+++ test/cpp/redis/t18_pipeline.cc
@@ -1,4 +1,4 @@
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-kernel/core/runtime.h>
 #include "integration_config.h"
 #include <array>