+++ test/cpp/redis/t20_auth.cc
@@ -1,4 +1,4 @@
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-redis/sync/redis_session.h>
 
 #include <galay/cpp/galay-kernel/core/runtime.h>