+++ test/cpp/redis/t25_rediss_pool_acquire_connected.cc
@@ -1,5 +1,5 @@
 #include <galay/cpp/galay-redis/async/conn_pool.h>
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-kernel/core/runtime.h>
 
 #include <atomic>