+++ test/cpp/redis/t5_command.cc
@@ -2,7 +2,7 @@
 #include <concepts>
 #include <utility>
 
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 
 using namespace galay::redis;