+++ test/cpp/redis/t6_pubsub.cc
@@ -7,7 +7,7 @@
 #include <mutex>
 #include <thread>
 
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-redis/async/topology_client.h>
 
 using namespace galay::kernel;