+++ test/cpp/redis/t8_cluster.cc
@@ -11,7 +11,7 @@
 #include <mutex>
 #include <string>
 
-#include <galay/cpp/galay-redis/async/redis_client.h>
+#include <galay/cpp/galay-redis/async/client.h>
 #include <galay/cpp/galay-redis/async/topology_client.h>
 
 using namespace galay::kernel;